refactor : GameRequestDto, MoveRequestDto 에서 도메인 객체를 생성해주던 to~ 메서드 제거
