+++ src/main/java/chess/controller/web/SpringWebChessRestController.java
@@ -1,12 +1,14 @@
 package chess.controller.web;
 
-import chess.controller.web.dto.GameIdDto;
+import chess.controller.web.dto.game.GameIdResponseDto;
 import chess.controller.web.dto.game.GameRequestDto;
 import chess.controller.web.dto.history.HistoryResponseDto;
 import chess.controller.web.dto.move.MoveRequestDto;
 import chess.controller.web.dto.piece.PieceResponseDto;
 import chess.controller.web.dto.score.ScoreResponseDto;
 import chess.controller.web.dto.state.StateResponseDto;
+import chess.domain.game.Game;
+import chess.domain.movecommand.MoveCommand;
 import chess.service.ChessService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,9 +26,12 @@ public class SpringWebChessRestController {
     }
 
     @PostMapping("")
-    public ResponseEntity<GameIdDto> saveGame(@RequestBody GameRequestDto gameRequestDto) {
-        GameIdDto gameIdDto = new GameIdDto(chessService.saveGame(gameRequestDto.toGame()));
-        return ResponseEntity.ok().body(gameIdDto);
+    public ResponseEntity<GameIdResponseDto> saveGame(@RequestBody GameRequestDto gameRequestDto) {
+        Game game = Game.of(gameRequestDto.getRoomName(),
+                gameRequestDto.getWhiteUsername(),
+                gameRequestDto.getBlackUsername());
+        GameIdResponseDto gameIdResponseDto = new GameIdResponseDto(chessService.saveGame(game));
+        return ResponseEntity.ok().body(gameIdResponseDto);
     }
 
     @GetMapping("/{id}/load")
@@ -56,6 +61,7 @@ public class SpringWebChessRestController {
 
     @PostMapping("/{id}/move")
     public ResponseEntity<HistoryResponseDto> move(@PathVariable Long id, @RequestBody MoveRequestDto moveRequestDto) {
-        return ResponseEntity.ok().body(chessService.move(moveRequestDto.toMoveCommand(), id));
+        MoveCommand moveCommand = new MoveCommand(moveRequestDto.getSource(), moveRequestDto.getTarget());
+        return ResponseEntity.ok().body(chessService.move(moveCommand, id));
     }
 }
