+++ src/main/java/chess/controller/web/dto/exception/ExceptionMessageResponseDto.java
@@ -1,4 +1,4 @@
-package chess.controller.web.dto;
+package chess.controller.web.dto.exception;
 
 public class ExceptionMessageResponseDto {
